flask-app/requirements.txt
flask
flask_sqlalchemy
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-flask
prometheus_client
structlog
requests
